Don't need to draw and save charts everytime buy/sell score is computed

+++ gammath_stocks_analysis.py
@@ -104,33 +104,34 @@ def get_ticker_hist_n_analysis(tsymbol):
         f.write(f'{price_signals}\n{rsi_signals}\n{bb_signals}\n{macd_signals}\n{mfi_signals}\n{stoch_slow_signals}\n{st_signals}\n{options_signals}\n{overall_buy_rec}\n{overall_sell_rec}')
         f.close()
 
-        #Draw the charts to view all at once as subplots
-        figure, axes = plt.subplots(nrows=6, figsize=(21, 19))
+        if not (path / f'{tsymbol}_charts.png').exists():
+            #Draw the charts to view all at once as subplots
+            figure, axes = plt.subplots(nrows=6, figsize=(21, 19))
 
-        sym_str = f'{tsymbol}'
+            sym_str = f'{tsymbol}'
 
-        plot_data1 = pd.DataFrame({sym_str: df.Close, 'Upper Band': ub, 'Middle Band': mb, 'Lower Band': lb})
-        plot_data2 = pd.DataFrame({'RSI': rsi})
-        #Don't need to draw the MACD histogram
-        plot_data3 = pd.DataFrame({'MACD': macd, 'MACD_SIGNAL': macd_signal})
-        plot_data4 = pd.DataFrame({'MFI': mfi})
-        plot_data5 = pd.DataFrame({'SLOWK': slowk, 'SLOWD': slowd})
-        plot_data6 = pd.DataFrame({sym_str: df.Close, 'Kalman Filter': state_means.flatten()})
+            plot_data1 = pd.DataFrame({sym_str: df.Close, 'Upper Band': ub, 'Middle Band': mb, 'Lower Band': lb})
+            plot_data2 = pd.DataFrame({'RSI': rsi})
+            #Don't need to draw the MACD histogram
+            plot_data3 = pd.DataFrame({'MACD': macd, 'MACD_SIGNAL': macd_signal})
+            plot_data4 = pd.DataFrame({'MFI': mfi})
+            plot_data5 = pd.DataFrame({'SLOWK': slowk, 'SLOWD': slowd})
+            plot_data6 = pd.DataFrame({sym_str: df.Close, 'Kalman Filter': state_means.flatten()})
 
-        plot_data1.plot(ax=axes[0],lw=1,title='Bollinger Bands')
-        plot_data2.plot(ax=axes[1],lw=1,title='Relative Strength Index')
-        axes[1].axhline(RSI_OVERBOUGHT_LEVEL,lw=1,ls='-',c='r')
-        axes[1].axhline(RSI_OVERSOLD_LEVEL,lw=1,ls='-',c='r')
-        plot_data3.plot(ax=axes[2],lw=1,title='Moving Average Convergence Divergence')
-        plot_data4.plot(ax=axes[3],lw=1,title='Money Flow Index')
-        axes[3].axhline(MFI_OVERBOUGHT_LEVEL,lw=1,ls='-',c='r')
-        axes[3].axhline(MFI_OVERSOLD_LEVEL,lw=1,ls='-',c='r')
-        plot_data5.plot(ax=axes[4],lw=1,title='Stochastic Slow')
-        axes[4].axhline(STOCH_OVERBOUGHT_LEVEL,lw=1,ls='-',c='r')
-        axes[4].axhline(STOCH_OVERSOLD_LEVEL,lw=1,ls='-',c='r')
-        plot_data6.plot(ax=axes[5], lw=1,title='Kalman Filter')
+            plot_data1.plot(ax=axes[0],lw=1,title='Bollinger Bands')
+            plot_data2.plot(ax=axes[1],lw=1,title='Relative Strength Index')
+            axes[1].axhline(RSI_OVERBOUGHT_LEVEL,lw=1,ls='-',c='r')
+            axes[1].axhline(RSI_OVERSOLD_LEVEL,lw=1,ls='-',c='r')
+            plot_data3.plot(ax=axes[2],lw=1,title='Moving Average Convergence Divergence')
+            plot_data4.plot(ax=axes[3],lw=1,title='Money Flow Index')
+            axes[3].axhline(MFI_OVERBOUGHT_LEVEL,lw=1,ls='-',c='r')
+            axes[3].axhline(MFI_OVERSOLD_LEVEL,lw=1,ls='-',c='r')
+            plot_data5.plot(ax=axes[4],lw=1,title='Stochastic Slow')
+            axes[4].axhline(STOCH_OVERBOUGHT_LEVEL,lw=1,ls='-',c='r')
+            axes[4].axhline(STOCH_OVERSOLD_LEVEL,lw=1,ls='-',c='r')
+            plot_data6.plot(ax=axes[5], lw=1,title='Kalman Filter')
 
-        plt.savefig(path / f'{tsymbol}_charts.png')
+            plt.savefig(path / f'{tsymbol}_charts.png')
     except:
         print('\nError while getting stock info for ', tsymbol, ': ', sys.exc_info()[0])
